requirements.txt
streamlit
ultralytics
opencv-python-headless
Pillow
roboflow
python-dotenv
